+++ src/BouncyCastle.Core/crypto/digests/Blake2bDigest.cs
@@ -290,7 +290,7 @@ namespace org.bouncycastle.crypto.digests
 					t1++;
 				}
 				compress(buffer, 0);
-				Arrays.fill(buffer, (byte)0); // clear buffer
+				Arrays.fill(buffer, 0); // clear buffer
 				buffer[0] = b;
 				bufferPos = 1;
 			}
@@ -333,7 +333,7 @@ namespace org.bouncycastle.crypto.digests
 					}
 					compress(buffer, 0);
 					bufferPos = 0;
-					Arrays.fill(buffer, (byte)0); // clear buffer
+					Arrays.fill(buffer, 0); // clear buffer
 				}
 				else
 				{
@@ -379,7 +379,7 @@ namespace org.bouncycastle.crypto.digests
 				t1++;
 			}
 			compress(buffer, 0);
-			Arrays.fill(buffer, (byte)0); // Holds eventually the key if input is null
+			Arrays.fill(buffer, 0); // Holds eventually the key if input is null
 			Arrays.fill(internalState, 0L);
 
 			for (int i = 0; i < chainValue.Length && (i * 8 < digestLength); i++)
@@ -415,7 +415,7 @@ namespace org.bouncycastle.crypto.digests
 			t0 = 0L;
 			t1 = 0L;
 			chainValue = null;
-			Arrays.fill(buffer, (byte)0);
+			Arrays.fill(buffer, 0);
 			if (key != null)
 			{
 				JavaSystem.arraycopy(key, 0, buffer, 0, key.Length);
@@ -512,8 +512,8 @@ namespace org.bouncycastle.crypto.digests
 		{
 			if (key != null)
 			{
-				Arrays.fill(key, (byte)0);
-				Arrays.fill(buffer, (byte)0);
+				Arrays.fill(key, 0);
+				Arrays.fill(buffer, 0);
 			}
 		}
 
@@ -525,7 +525,7 @@ namespace org.bouncycastle.crypto.digests
 		{
 			if (salt != null)
 			{
-				Arrays.fill(salt, (byte)0);
+				Arrays.fill(salt, 0);
 			}
 		}
 	}
