+++ src/BouncyCastle.Core/crypto/digests/SHA384Digest.cs
@@ -80,13 +80,13 @@
 			 * The first 64 bits of the fractional parts of the square roots
 			 * of the 9th through 16th prime numbers
 			 */
-			H1 = unchecked((ulong)0xcbbb9d5dc1059ed8Ul);
+			H1 = unchecked(0xcbbb9d5dc1059ed8Ul);
 			H2 = 0x629a292a367cd507l;
-			H3 = unchecked((ulong)0x9159015a3070dd17Ul);
+			H3 = unchecked(0x9159015a3070dd17Ul);
 			H4 = 0x152fecd8f70e5939Ul;
 			H5 = 0x67332667ffc00b31Ul;
-			H6 = unchecked((ulong)0x8eb44a8768581511Ul);
-			H7 = unchecked((ulong)0xdb0c2e0d64f98fa7Ul);
+			H6 = unchecked(0x8eb44a8768581511Ul);
+			H7 = unchecked(0xdb0c2e0d64f98fa7Ul);
 			H8 = 0x47b5481dbefa4fa4Ul;
 		}
